+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Not applicable
+- Partial support for CDMA and WCDMA towers when using getAllCellInfo() API.
 
 ### Changed
 - Check for unknown values in fields in the cell IDs returned by getAllCellInfo();